Create Exporter class for export methods
